Factor out `_start_bg_tasks` from farmer.
